refactor StreamFileEmChunks

+++ @junioranheu-utils-package/Fixtures/Convert.cs
@@ -77,7 +77,7 @@ namespace junioranheu_utils_package.Fixtures
         /// <summary>
         /// Converter caminho de um arquivo para stream;
         /// </summary>
-        public static async Task<Stream?> ConverterPathParaStream(string path, int? chunkSize = 4096)
+        public static async Task<Stream?> ConverterPathParaStream(string path, long? chunkSize = 4096)
         {
             if (string.IsNullOrEmpty(path))
             {
@@ -89,7 +89,7 @@ namespace junioranheu_utils_package.Fixtures
                 return null;
             }
 
-            return await Task.FromResult(new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.Read, chunkSize.GetValueOrDefault(), FileOptions.Asynchronous));
+            return await Task.FromResult(new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.Read, (int)chunkSize.GetValueOrDefault(), FileOptions.Asynchronous));
         }
 
         /// <summary>
